refactor: ページの部分要素をテンプレートに分離
