+++ src/templates/index.ts
@@ -1,0 +1,6 @@
+import { HistoryTemplate } from "./history"
+import { IntroTemplate } from "./intro"
+import { ProfileTemplate } from "./profile"
+import { SkillTemplate } from "./skill"
+import { TimelineTemplate } from "./timeline"
+export { HistoryTemplate, IntroTemplate, ProfileTemplate, SkillTemplate, TimelineTemplate }
